Moved CanvasImpl (as well as its associated classes) to NUI.
