+++ engine/src/main/java/org/terasology/engine/subsystem/headless/HeadlessGraphics.java
@@ -52,7 +52,7 @@ import org.terasology.rendering.assets.texture.Texture;
 import org.terasology.rendering.assets.texture.TextureData;
 import org.terasology.rendering.assets.texture.subtexture.Subtexture;
 import org.terasology.rendering.assets.texture.subtexture.SubtextureData;
-import org.terasology.rendering.nui.internal.CanvasRenderer;
+import org.terasology.nui.canvas.CanvasRenderer;
 
 public class HeadlessGraphics implements EngineSubsystem {
