+++ engine/src/main/java/org/terasology/rendering/nui/CanvasUtility.java
@@ -17,7 +17,6 @@ package org.terasology.rendering.nui;
 
 import org.terasology.assets.ResourceUrn;
 import org.terasology.math.geom.BaseVector2i;
-import org.terasology.math.geom.ImmutableVector2i;
 import org.terasology.math.geom.Quat4f;
 import org.terasology.math.geom.Rect2i;
 import org.terasology.math.geom.Vector3f;
@@ -26,7 +25,7 @@ import org.terasology.nui.SubRegion;
 import org.terasology.rendering.assets.material.Material;
 import org.terasology.rendering.assets.mesh.Mesh;
 import org.terasology.rendering.assets.texture.Texture;
-import org.terasology.rendering.nui.internal.CanvasImpl;
+import org.terasology.rendering.nui.internal.TerasologyCanvasImpl;
 
 // HACK: This whole class was created in order to provide access to internal, implementation specific methods.
 
@@ -40,40 +39,40 @@ public final class CanvasUtility {
     public static void drawMesh(Canvas canvas, Mesh mesh, Texture texture, Rect2i region, Quat4f rotation, Vector3f offset, float scale) {
         // TODO: Find a way to abstractly implement drawMesh in NUI
 
-        if (!(canvas instanceof CanvasImpl)) {
+        if (!(canvas instanceof TerasologyCanvasImpl)) {
             throw new UnsupportedOperationException("Drawing meshes is only supported using Terasology's internal renderer.");
         }
 
-        ((CanvasImpl) canvas).drawMesh(mesh, texture, region, rotation, offset, scale);
+        ((TerasologyCanvasImpl) canvas).drawMesh(mesh, texture, region, rotation, offset, scale);
     }
 
     public static void drawMesh(Canvas canvas, Mesh mesh, Material material, Rect2i region, Quat4f rotation, Vector3f offset, float scale) {
         // TODO: Find a way to abstractly implement drawMesh in NUI
 
-        if (!(canvas instanceof CanvasImpl)) {
+        if (!(canvas instanceof TerasologyCanvasImpl)) {
             throw new UnsupportedOperationException("Drawing meshes is only supported using Terasology's internal renderer.");
         }
 
-        ((CanvasImpl) canvas).drawMesh(mesh, material, region, rotation, offset, scale);
+        ((TerasologyCanvasImpl) canvas).drawMesh(mesh, material, region, rotation, offset, scale);
     }
 
     public static void drawMaterial(Canvas canvas, Material material, Rect2i region) {
         // TODO: Find a way to abstractly implement drawMaterial in NUI
 
-        if (!(canvas instanceof CanvasImpl)) {
+        if (!(canvas instanceof TerasologyCanvasImpl)) {
             throw new UnsupportedOperationException("Drawing materials is only supported using Terasology's internal renderer.");
         }
 
-        ((CanvasImpl) canvas).drawMaterial(material, region);
+        ((TerasologyCanvasImpl) canvas).drawMaterial(material, region);
     }
 
     public static SubRegion subRegionFBO(Canvas canvas, ResourceUrn uri, BaseVector2i size) {
         // TODO: Find a way to abstractly implement subRegionFBO in NUI
 
-        if (!(canvas instanceof CanvasImpl)) {
+        if (!(canvas instanceof TerasologyCanvasImpl)) {
             throw new UnsupportedOperationException("Used FBO sub-regions is only supported using Terasology's internal renderer.");
         }
 
-        return ((CanvasImpl) canvas).subRegionFBO(uri, size);
+        return ((TerasologyCanvasImpl) canvas).subRegionFBO(uri, size);
     }
 }
